Rename profiles routes file
